Moving compression plugins to components directory
- flyby: removing obsolete #include files
