+++ components/parcel_plugins/binary_filter/bzip2/include/hpx/include/compression_bzip2.hpp
@@ -1,0 +1,10 @@
+//  Copyright (c) 2007-2022 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#pragma once
+
+#include <hpx/config.hpp>
+#include <hpx/binary_filter/bzip2_serialization_filter.hpp>
