+++ components/parcel_plugins/binary_filter/snappy/include/hpx/include/compression_snappy.hpp
@@ -1,0 +1,10 @@
+//  Copyright (c) 2007-2022 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#pragma once
+
+#include <hpx/config.hpp>
+#include <hpx/binary_filter/snappy_serialization_filter.hpp>
