+++ libs/full/parcelset/tests/unit/CMakeLists.txt
@@ -15,15 +15,6 @@ set(tests put_parcels set_parcel_write_handler)
 set(put_parcels_PARAMETERS LOCALITIES 2)
 set(set_parcel_write_handler_PARAMETERS LOCALITIES 2)
 
-if(HPX_WITH_COMPRESSION_BZIP2
-   OR HPX_WITH_COMPRESSION_ZLIB
-   OR HPX_WITH_COMPRESSION_SNAPPY
-)
-  set(tests ${tests} put_parcels_with_compression)
-  set(put_parcels_with_compression_PARAMETERS LOCALITIES 2)
-  set(put_parcels_with_compression_FLAGS DEPENDENCIES iostreams_component)
-endif()
-
 foreach(test ${tests})
   set(sources ${test}.cpp)
