+++ plugins/CMakeLists.txt
@@ -1,10 +1,10 @@
-# Copyright (c) 2007-2021 Hartmut Kaiser
+# Copyright (c) 2007-2022 Hartmut Kaiser
 #
 # SPDX-License-Identifier: BSL-1.0
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-set(plugin_categories binary_filter parcelport)
+set(plugin_categories parcelport)
 
 foreach(category ${plugin_categories})
   add_hpx_pseudo_target(plugins.${category})
